fix(category): 잼민이 요청 잘못된 오류 수정

+++ src/main/java/com/example/chicken/domain/product/dto/CategoryCountResponseDto.java
@@ -7,5 +7,5 @@ public record CategoryCountResponseDto(Long id,
                                        String name,
                                        LocalDateTime createdAt,
                                        LocalDateTime modifiedAt,
-                                       Integer productCount) {
+                                       Long productCount) {
 }
